backend/recipe-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.7.1
python-dotenv==1.0.0
requests==2.32.3
